DisplayController/requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-bme680==3.7.12
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-sgp30==3.0.11
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
asyncio==3.4.3
binho-host-adapter==0.1.6
board==1.0
colorama==0.4.6
coverage==7.8.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-mock==3.14.0
pyusb==1.3.1
robotframework==7.2.2
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus2==0.5.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
websockets==15.0.1
